--- DashboardV2/src/components/WireShark.tsx
@@ -1,10 +1,0 @@
-
-const WireShark = () => {
-  return (
-    <>
-      <h1 style={{textAlign:'center'}}>เดะหาอะไรมาแทนหน้านี้ อย่าเพิ่งรีบนะพี่ชาย</h1>
-    </>
-  )
-}
-
-export default WireShark
